requirements.txt
flask==3.1.0
pandas==2.2.3
openpyxl==3.1.5
sentence-transformers==3.4.1
chromadb==0.6.3
groq==0.15.0
python-dotenv==1.0.1
tqdm==4.67.1